refactor(transforms): migrate transforms to lib/ and auto-discovery
