+++ src/transforms/index.js
@@ -1,18 +1,21 @@
-const parseFrontmatter = require('./parse-frontmatter');
-const extractExamples = require('./extract-examples');
-const rebuildFrontmatter = require('./rebuild-frontmatter');
-const skillMetadata = require('./skill-metadata');
-const agentStub = require('./agent-stub');
-const skillDiscoveryStub = require('./skill-discovery-stub');
+'use strict';
 
-const transforms = {
-  'parse-frontmatter': parseFrontmatter,
-  'extract-examples': extractExamples,
-  'rebuild-frontmatter': rebuildFrontmatter,
-  'skill-metadata': skillMetadata,
-  'agent-stub': agentStub,
-  'skill-discovery-stub': skillDiscoveryStub,
-};
+const path = require('node:path');
+const { discover } = require('../lib/discovery');
+
+const TRANSFORMS_DIR = path.resolve(__dirname);
+
+const entries = discover({
+  dir: TRANSFORMS_DIR,
+  pattern: '*.js',
+  identity: (filepath) => path.basename(filepath, '.js'),
+  validate: (entry) => entry.id !== 'index',
+});
+
+const transforms = Object.create(null);
+for (const entry of entries) {
+  transforms[entry.id] = require(path.join(TRANSFORMS_DIR, `${entry.id}.js`));
+}
 
 /**
  * Resolve a transform name to its function.
